chore(tests): Restructure API test files for clarity
Moved `AppTest` class to a separate file `test_app.py` for better
organization. Simplified imports by removing unused dependencies in
`test_access_lists.py`.
